Continuing on modularization (fetcher)
